[rename] MainPage 이름 변경
